Move RenderDebouncer and ScreenDprMonitor into ui
